adding spring easing from react-motion for animation
The "right way" to do animation in React seems to be to save some state in a
component. That state is then used to animate attributes when new props are
set. This keeps everything consistent.

R2D3 does something similar. It uses an internal ZENO function as an easing
function. The target state is requested, and internal state is animated
towards the target state using ZENO. In the case of the tree, ZENO is used on
the scroll state and the code fires its own fake scroll events until the
internal state hits the target.

I started reimplementing this general idea in this clone, but after reading
the react-motion source, realized that I was just going to reimplement its
general approach (and worse). So now we use react-motion's StaggeredMotion as
an easy way to set easing on each sample's progress. We are _not_ using
StaggeredMotion for the linked springs effect that is in the examples. In
tests, it was hard to get the separation I wanted for each sample.