factor out tree sub-components into their own modules
A tree is composed of leaves, links, and paths to each leaf. These
subcomponents were defined within the tree component to leverage
closure over variables in tree. But tree became hard to read.

This commit factors all of those components out of the tree. The
necessary variables are passed via arguments. This style looks less
clean in the subcomponents, but tree looks a lot cleaner.